Split truncate table driver specs
